% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClient.R
\name{CheckAWSKeys}
\alias{CheckAWSKeys}
\title{Helper function to check AWS Keys}
\usage{
CheckAWSKeys()
}
\value{
A logical indicating whether AWS Keys were found as environment variables.
}
\description{
Checks for the existence of environment variables AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.
}
