% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisableHIT.R
\name{DisableHIT}
\alias{DisableHIT}
\alias{DisposeHIT}
\alias{ExpireHIT}
\alias{DeleteHIT}
\alias{disablehit}
\alias{disposehit}
\alias{deletehit}
\title{Disable HIT}
\usage{
DisableHIT(hit = NULL, hit.type = NULL, annotation = NULL,
  approve.pending.assignments = FALSE, delete.hit = FALSE,
  verbose = TRUE)
}
\arguments{
\item{hit}{A character string containing a HITId or a vector of character
strings containing multiple HITIds. Must specify \code{hit} xor
\code{hit.type} xor \code{annotation}.}

\item{hit.type}{An optional character string containing a HITTypeId (or a
vector of HITTypeIds). Must specify \code{hit} xor \code{hit.type} xor
\code{annotation}.}

\item{annotation}{An optional character string specifying the value of the
\code{RequesterAnnotation} field for a batch of HITs. This can be used to
disable all HITs from a \dQuote{batch} created in the online Requester User
Interface (RUI). To use a batch ID, the batch must be written in a character
string of the form \dQuote{BatchId:78382;}, where \dQuote{73832} is the
batch ID shown in the RUI. Must specify \code{hit} xor \code{hit.type} xor
\code{annotation}.}

\item{approve.pending.assignments}{A logical indicating whether the pending
assignments should be approved when the HIT is disabled.}

\item{delete.hit}{A logical indicating whether to try deleting the HIT
when the HIT is disabled. This will also delete the assignment data. Note
that the HIT can only be deleted if it has a "Reviewing" or "Reviewable"
status. In other words, you can't delete a HIT if someone is currently working
on an assignment for it.}
}
\value{
A data frame containing a list of HITs and whether the request to
disable each of them was valid.
}
\description{
This function will allow you to expire a HIT early, which means it will no
longer be available for new workers to accept. Optionally, when disabling the
HIT you can approve all pending assignments and you can also try to delete
the HIT. Be careful when deleting a HIT: this will also delete the assignment
data!
}
\details{
If you disable a HIT while workers are still working on an assignment, they
will still be able to complete their task.

\code{DisposeHIT()}, \code{ExpireHIT()}, \code{DeleteHIT()}, \code{disablehit()},
\code{disposehit()}, \code{expirehit()}, \code{deletehit()} are aliases.
}
\examples{

\dontrun{
# Disable a single HIT
b <- GenerateExternalQuestion("http://www.example.com/","400")
hit1 <- CreateHIT(hit.type="2FFNCWYB49F9BBJWA4SJUNST5OFSOW",
                  expiration = seconds(days = 1),
                  question=b$string)
DisableHIT(hit = hit1$HITId)

# Disable all HITs of a given HITType
DisableHIT(hit.type = hit1$HITTypeId)

# Disable all HITs of a given batch from the RUI
DisableHIT(annotation="BatchId:78382;")
}

}
\references{
\href{https://docs.aws.amazon.com/AWSMechTurk/latest/AWSMturkAPI/ApiReference_UpdateExpirationForHITOperation.html}{API Reference: Update Expiration for HIT}
\href{https://docs.aws.amazon.com/AWSMechTurk/latest/AWSMturkAPI/ApiReference_DeleteHITOperation.html}{API Reference: Delete HIT}
}
\author{
Tyler Burleigh, Thomas J. Leeper
}
\keyword{HITs}
