% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptydf.R
\name{emptydf}
\alias{emptydf}
\title{Helper function that creates an empty data.frame}
\usage{
emptydf(nrow, ncol, names)
}
\arguments{
\item{nrow}{Number of rows}

\item{nrow}{Number of columns}

\item{nrow}{Number of names of the columns}
}
\description{
Helper function that creates an empty data.frame
}
