% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHIT.R
\name{GetHIT}
\alias{GetHIT}
\alias{gethit}
\alias{hit}
\title{Get HIT}
\usage{
GetHIT(hit, return.hit.dataframe = TRUE, return.qual.dataframe = TRUE,
  verbose = TRUE)
}
\arguments{
\item{hit}{A character string specifying the HITId of the HIT to be retrieved.}

\item{return.hit.dataframe}{A logical indicating whether the data frame of
HITs should be returned. Default is \code{TRUE}.}

\item{return.qual.dataframe}{A logical indicating whether the list of each
HIT's QualificationRequirements (stored as data frames in that list) should
be returned. Default is \code{TRUE}.}
}
\description{
Retrieve various details of a HIT as a data frame.
}
\details{
\code{GetHIT} retrieves characteristics of a HIT. \code{HITStatus} is a
wrapper that retrieves the Number of Assignments Pending, Number of
Assignments Available, Number of Assignments Completed for the HIT(s), which
is helpful for checking on the progress of currently available HITs.

\code{gethit()} and \code{hit()} are aliases for \code{GetHIT}.
}
\references{
\href{https://docs.aws.amazon.com/AWSMechTurk/latest/AWSMturkAPI/ApiReference_GetHITOperation.html}{API Reference}
}
\author{
Tyler Burleigh, Thomas J. Leeper
}
\keyword{HITs}
