% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchHITs.R
\name{SearchHITs}
\alias{SearchHITs}
\alias{searchhits}
\alias{ListHITs}
\alias{listhits}
\title{Search your HITs}
\usage{
SearchHITs(return.all = TRUE, results = as.integer(10),
  pagetoken = NULL, sandbox = TRUE, profile = "default",
  verbose = TRUE)
}
\arguments{
\item{return.all}{A logical indicating whether all HITs (as opposed to a
small subjset) should be returned. Default is \code{TRUE}.}

\item{results}{An optional character string indicating how many search
results should be returned by each request, between 1 and 100. Most users
can ignore this.}

\item{pagetoken}{An optional character string indicating which page of
search results should be returned. Most users can ignore this.}
}
\value{
A list containing data frames of HITs and Qualification Requirements
}
\description{
Search for your HITs and return those HITs as R objects.
}
\details{
Retrieve your current HITs (and, optionally, characteristics thereof).

\code{searchhits()}, \code{ListHITs()}, and \code{listhits()} are aliases
}
\examples{

\dontrun{
SearchHITs()
}

}
\references{
\href{https://docs.aws.amazon.com/AWSMechTurk/latest/AWSMturkAPI/ApiReference_ListHITsOperation.html}{API
Reference}
}
\author{
Tyler Burleigh, Thomas J. Leeper
}
\keyword{HITs}
