% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seconds.R
\name{seconds}
\alias{seconds}
\title{Convert arbitrary times to seconds}
\usage{
seconds(days = NULL, hours = NULL, minutes = NULL, seconds = NULL)
}
\arguments{
\item{days}{An optional number of days.}

\item{hours}{An optional number of hours.}

\item{minutes}{An optional number of minutes.}

\item{seconds}{An optional number of seconds.}
}
\value{
An integer equal to the requested amount of time in seconds.
}
\description{
A convenience function to convert arbitrary numbers of days, hours, minutes,
and/or seconds into seconds.
}
\details{
A convenience function to convert arbitrary numbers of days, hours, minutes,
and/or seconds into seconds. For example, to be used in setting a HIT
expiration time. MTurk only accepts times (e.g., for HIT expirations, or the
duration of assignments) in seconds. This function returns an integer value
equal to the number of seconds of the input, and can be used atomically
within other MTurkR calls (e.g., \code{\link{CreateHIT}}).
}
\author{
Thomas J. Leeper
}
