% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameAssignment}
\alias{ToDataFrameAssignment}
\title{ToDataFrameAssignment}
\usage{
ToDataFrameAssignment(assignment)
}
\arguments{
\item{assignment}{assignment}
}
\value{
A list of Data frames, for assignment information and answers
}
\description{
Get a list of assignment and answer information for an assignment
}
