% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameBonusPayments}
\alias{ToDataFrameBonusPayments}
\title{ToDataFrameBonusPayments}
\usage{
ToDataFrameBonusPayments(bonuses)
}
\arguments{
\item{bonuses}{bonuses}
}
\value{
A Data frame of Bonus payment information
}
\description{
ToDataFrameBonusPayments
}
