% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameHITs}
\alias{ToDataFrameHITs}
\title{ToDataFrameHITs}
\usage{
ToDataFrameHITs(hits)
}
\arguments{
\item{hits}{hits}
}
\value{
A data frame of information on HITs, one per row.
}
\description{
Convert a list of HITs to a data frame
}
