% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameQualificationRequirements}
\alias{ToDataFrameQualificationRequirements}
\title{ToDataFrameQualificationRequirements}
\usage{
ToDataFrameQualificationRequirements(hits)
}
\arguments{
\item{hits}{hits}
}
\description{
ToDataFrameQualificationRequirements
}
