% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameQualifications}
\alias{ToDataFrameQualifications}
\title{ToDataFrameQualifications}
\usage{
ToDataFrameQualifications(quals)
}
\arguments{
\item{quals}{qualifications}
}
\value{
A Data frame of qualification information
}
\description{
ToDataFrameQualifications
}
