% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameQualificationTypes}
\alias{ToDataFrameQualificationTypes}
\title{ToDataFrameQualificationTypes}
\usage{
ToDataFrameQualificationTypes(quals)
}
\arguments{
\item{quals}{qualifications}
}
\value{
A Data frame of Qualification Types
}
\description{
ToDataFrameQualificationTypes
}
