% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameQuestionFormAnswers}
\alias{ToDataFrameQuestionFormAnswers}
\title{ToDataFrameQuestionFormAnswers}
\usage{
ToDataFrameQuestionFormAnswers(assignment, answers)
}
\description{
ToDataFrameQuestionFormAnswers
}
