% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameReviewableHITs}
\alias{ToDataFrameReviewableHITs}
\title{ToDataFrameReviewableHITs}
\usage{
ToDataFrameReviewableHITs(hits)
}
\arguments{
\item{hits}{hits}
}
\value{
A Data frame of reviewable HIT information
}
\description{
ToDataFrameReviewableHITs
}
