% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameReviewResults}
\alias{ToDataFrameReviewResults}
\title{ToDataFrameReviewResults}
\usage{
ToDataFrameReviewResults(results)
}
\arguments{
\item{results}{results}
}
\value{
A list of Data frames of Assignment Reviews/Actions and HIT Reviews/Actions.
}
\description{
ToDataFrameReviewResults
}
