% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameWorkerBlock}
\alias{ToDataFrameWorkerBlock}
\title{ToDataFrameWorkerBlock}
\usage{
ToDataFrameWorkerBlock(workers)
}
\arguments{
\item{workers}{workers}
}
\value{
A Data frame of blocked workers
}
\description{
ToDataFrameWorkerBlock
}
