% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHIT.R
\name{GetHIT}
\alias{GetHIT}
\alias{gethit}
\alias{hit}
\alias{HITStatus}
\alias{status}
\title{Get HIT}
\usage{
GetHIT(hit, verbose = TRUE)
}
\arguments{
\item{hit}{A character string specifying the HITId of the HIT to be retrieved.}
}
\description{
Retrieve various details of a HIT as a data frame.
}
\details{
\code{GetHIT} retrieves characteristics of a HIT. \code{HITStatus} is a
wrapper that retrieves the Number of Assignments Pending, Number of
Assignments Available, Number of Assignments Completed for the HIT(s), which
is helpful for checking on the progress of currently available HITs.

\code{gethit()} and \code{hit()} are aliases for \code{GetHIT}.
\code{status()} is an alias for \code{HITStatus}.
}
\references{
\href{https://docs.aws.amazon.com/AWSMechTurk/latest/AWSMturkAPI/ApiReference_GetHITOperation.html}{API Reference}
}
\author{
Tyler Burleigh, Thomas J. Leeper
}
\keyword{HITs}
