% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameAssignment}
\alias{ToDataFrameAssignment}
\title{ToDataFrameAssignment}
\usage{
ToDataFrameAssignment(assignment)
}
\arguments{
\item{assignment}{assignment}
}
\description{
ToDataFrameAssignment
}
