% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameQualificationRequests}
\alias{ToDataFrameQualificationRequests}
\title{ToDataFrameQualificationRequests}
\usage{
ToDataFrameQualificationRequests(requests)
}
\arguments{
\item{requests}{requests}
}
\value{
A Data frame of Qualification Request information
}
\description{
ToDataFrameQualificationRequests
}
