% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToDataFrame.R
\name{ToDataFrameQuestionFormAnswers}
\alias{ToDataFrameQuestionFormAnswers}
\title{ToDataFrameQuestionFormAnswers}
\usage{
ToDataFrameQuestionFormAnswers(assignment, answers)
}
\arguments{
\item{assignment}{assignment}

\item{answers}{answers}
}
\value{
A Data frame of Answer information for the assignment
}
\description{
ToDataFrameQuestionFormAnswers
}
